debug

+++ examples/xdp/syn_pad.bpf.c
@@ -89,14 +89,14 @@ static __always_inline void set_tcp_csum(struct iphdr *iph, struct tcphdr *tcph)
   //initialize checksum to 0
   tcph->check = 0;
   while (tcpLen > 1) {
-	  sum += * (unsigned short *)tcph++;
+	  sum += * (unsigned long *)tcph++;
 	  tcpLen -= 2;
   }
 
 	//if any bytes left, pad the bytes and add
   if(tcpLen > 0) {
     //printf("+++++++++++padding, %dn", tcpLen);
-    sum += ((*(unsigned short *)tcph)&bpf_htons(0xFF00));
+    sum += ((*(unsigned long *)tcph)&bpf_htons(0xFF00));
   }
 
   //Fold 32-bit sum to 16 bits: add carrier to result
